Use a single Azure topic for all our async notifications
Managing the machinery for three topics is too annoying, and it's easier and quicker and generally better to do this than set up all the local + deployed resources to make N topics work.

Instead, we use the `EventType` of a notification to indicate, well, the type of the event (e.g. `parsed-portfolio`, `created-report`, etc)

I'll send out the associated Terraform changes momentarily