Don't create a new Azure Container Apps Job (ACAJ) every time
It turns out that the way Azure intends you to use Container Apps Jobs is to create a single 'job' and then invoke it (optionally changing the command/image/etc).

'Jobs' are more like 'templates', and are fairly heavy to create. This PR changes our config to just run things against and existing ACAJ, which makes everything much faster.