Disable CORS + request/response logs
This PR disables some of our noiser logs, in particular CORS + request/response logging. As noted in the code, most of this info is provided by our cloud environment when deployed, and when local, these things are generally just noisy. They can be selectively turned on as debugging requires.

Fixes #67